+++ js/utils/events.js
@@ -13,10 +13,24 @@ const EVENTS = {
 	DOMredirect: 'dom:redirect',
 	mediachange: 'mediachange',
 	moduleRegistered: 'moduleRegistered',
-	mouseEnter: 'mouseenter',
-	mouseLeave: 'mouseleave',
+	mouseenter: 'mouseenter',
+	mouseleave: 'mouseleave',
+	mouseover: 'mouseover',
+	mouseout: 'mouseout',
+	mouseup: 'mouseup',
+	mousedown: 'mousedown',
 	resize: 'resize',
-	scroll: 'scroll'
+	scroll: 'scroll',
+	click: 'click',
+	dblclick: 'dblclick',
+	change: 'change',
+	input: 'input',
+	blur: 'blur',
+	submit: 'submit',
+	reset: 'reset',
+	keyup: 'keyup',
+	keydown: 'keydown',
+	keypress: 'keypress'
 };
 
 // @INSERTPOINT :: @ref: js-events